Add files via upload
The bookmark feature in Power BI is a powerful tool that allows users to capture and save the current state of a report page, including the filters, slicers, and visual elements' configurations. I've used such an amazing option where you can just click "CTRL+click the PIC" to go the first slide or page.